refactor: promote deriveProjectName/buildDockerCpArgs to core/lib/docker
