+++ run/src/main.test.ts
@@ -87,7 +87,7 @@ describe("readKnownBlockedRules", () => {
 // Regression guard for the concurrent-step container/network collision:
 // both must always include "-p" + the project name, or Compose falls back
 // to an implicit, directory-derived project name shared by every
-// concurrent `run` step in the job (see lib/container.ts's
+// concurrent `run` step in the job (see core/lib/docker/container.ts's
 // deriveProjectName).
 // -----------------------------------------------------------------------
